requirements.txt
django
pillow

django-jazzmin

whitenoise
django-brotli
gunicorn

djangorestframework
drf_yasg2

django-import-export

datapackage